feat: convert tonbo_record into derive macro
