+++ tests/01-simple-record.rs
@@ -1,8 +1,8 @@
-use tonbo_macros::tonbo_record;
+use tonbo_macros::Record;
 
-#[tonbo_record]
+#[derive(Record, Debug)]
 pub struct User {
-    #[primary_key]
+    #[record(primary_key)]
     name: String,
     email: Option<String>,
     age: u8,
