+++ tonbo_macros/Cargo.toml
@@ -7,6 +7,7 @@ name = "tonbo_macros"
 version = "0.1.0"
 
 [dependencies]
+darling = "0.20"
 proc-macro2 = "1"
 quote = "1"
 syn = "2"
