+++ tonbo_macros/src/keys.rs
@@ -1,6 +1,7 @@
 use proc_macro2::{Ident, TokenStream};
 use syn::Type;
 
+#[derive(Clone)]
 pub(crate) struct PrimaryKey {
     pub(crate) name: Ident,
     pub(crate) base_ty: Type,
